notification-backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
reportlab>=3.5,<4.0
matplotlib>=3.4,<4.0
django-celery-beat>=2.2,<3.0
celery>=5.1,<6.0
pandas>=1.2,<2.0
django-filter>=2.4,<3.0
```